Remove cuttlefish warning about message_size_limit
The setting `message_size_limit` has been deprecated for long but
still when one runs `vernemq config generate` or starts the broker
without a configuration file, it warns about this setting even when is
not set. The setting has been deprecated for long and is not used
anywhere, so we should remove it from the schema file.